refactoring to follow the  format for importing
