+++ src/prompts/index.ts
@@ -1,9 +1,9 @@
 const { MultiSelect }  = require('enquirer');
-import { nginxPrompt, nginxHealthPrompt } from './nginx/nginx';
-import { apachePrompt, apacheHealthPrompt } from './Apache/apache';
-import { postgresPrompt, postgresCredentialsPrompt } from './Postgres/postgres';
-import { mongoPrompt, mongoCredentialsPrompt } from './MongoDB/mongodb';
-import { hostPrompt } from './Host/host';
+import { nginxPrompt, nginxHealthPrompt } from './Nginx';
+import { apachePrompt, apacheHealthPrompt } from './Apache';
+import { postgresPrompt, postgresCredentialsPrompt } from './Postgres';
+import { mongoPrompt, mongoCredentialsPrompt } from './MongoDB';
+import { hostPrompt } from './Host';
 
 const servicesToMonitor: any = new MultiSelect({
   name: 'sources',
